+++ src/news/models/news.interface.ts
